requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
google-generativeai
python-multipart
streamlit
requests
chardet
chromadb